--- src/main/java/com/charmroom/charmroom/repository/BoardReposiory.java
@@ -1,7 +1,0 @@
-package com.charmroom.charmroom.repository;
-
-import com.charmroom.charmroom.entity.Board;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BoardReposiory extends JpaRepository<Board, Integer> {
-}
