+++ src/main/java/com/charmroom/charmroom/service/MarketService.java
@@ -1,33 +1,73 @@
 package com.charmroom.charmroom.service;
 
+import com.charmroom.charmroom.dto.business.MarketDto;
 import com.charmroom.charmroom.entity.Article;
+import com.charmroom.charmroom.entity.Attachment;
+import com.charmroom.charmroom.entity.Board;
 import com.charmroom.charmroom.entity.Market;
+import com.charmroom.charmroom.entity.User;
 import com.charmroom.charmroom.entity.enums.MarketArticleState;
 import com.charmroom.charmroom.exception.BusinessLogicError;
 import com.charmroom.charmroom.exception.BusinessLogicException;
 import com.charmroom.charmroom.repository.ArticleRepository;
+import com.charmroom.charmroom.repository.AttachmentRepository;
+import com.charmroom.charmroom.repository.BoardRepository;
 import com.charmroom.charmroom.repository.MarketRepository;
+import com.charmroom.charmroom.repository.UserRepository;
+import com.charmroom.charmroom.util.CharmroomUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class MarketService {
     private final MarketRepository marketRepository;
     private final ArticleRepository articleRepository;
+    private final UserRepository userRepository;
+    private final BoardRepository boardRepository;
+    private final CharmroomUtil.Upload uploadUtils;
+    private final AttachmentRepository attachmentRepository;
 
-    public Market create(Article article, Integer price, String tag, MarketArticleState state) {
+    public Market create(MarketDto marketDto, List<MultipartFile> files) {
+        User user = userRepository.findByUsername(marketDto.getUsername()).orElseThrow(() ->
+                new BusinessLogicException(BusinessLogicError.NOTFOUND_USER));
+
+        Board board = boardRepository.findById(marketDto.getBoardId()).orElseThrow(() -> new BusinessLogicException(BusinessLogicError.NOTFOUND_BOARD));
+
+        Article article = Article.builder()
+                .user(user)
+                .board(board)
+                .title(marketDto.getTitle())
+                .body(marketDto.getBody())
+                .build();
+
+        for (MultipartFile file : files) {
+            Attachment attachment = uploadUtils.buildAttachment(file, article);
+            Attachment saved = attachmentRepository.save(attachment);
+            article.getAttachmentList().add(saved);
+        }
+
+        Article savedArticle = articleRepository.save(article);
+
         Market market = Market.builder()
-                .article(article)
-                .price(price)
-                .tag(tag)
-                .state(state)
+                .article(savedArticle)
+                .price(marketDto.getPrice())
+                .state(marketDto.getState())
+                .tag(marketDto.getTag())
                 .build();
 
         return marketRepository.save(market);
+    }
+
+    public Market create(MarketDto marketDto) {
+        return create(marketDto, new ArrayList<>());
     }
 
     public Page<Market> getAllMarketsByPageable(Pageable pageable) {
@@ -70,8 +110,7 @@ public class MarketService {
         Market market = marketRepository.findById(marketId).orElseThrow(() ->
                 new BusinessLogicException(BusinessLogicError.NOTFOUND_ARTICLE, "marketId: " + marketId));
 
-        Article article = market.getArticle();
-        articleRepository.delete(article);
         marketRepository.delete(market);
+        articleRepository.delete(market.getArticle());
     }
 }
